+++ src/main/java/dev/xkmc/lasertransport/content/capability/generic/HandlerWrapper.java
@@ -2,9 +2,9 @@ package dev.xkmc.lasertransport.content.capability.generic;
 
 public interface HandlerWrapper {
 
-	int insert(GenericHolder token, boolean simulate);
+	long insert(GenericHolder token, boolean simulate);
 
-	GenericHolder extract(int slot, int max, boolean b);
+	GenericHolder extract(int slot, long max, boolean b);
 
 	int getSize();
